Apple: Fix debuginfo of deoptimized files.
The -g flag was set in compileroptimization variable, so when a file
was deoptimized this flag was never set for those flags. In the Windows
BFF template we set the debug flags in the ExtraOptions, so I moved it
there since we always want to have the debug information if it's turned
on.